Test: 구입 수량 테스트 이전
